+++ src/main.py
@@ -6,7 +6,7 @@ import os
 import argparse
 from utils.yaml_utils import init_yaml_representer
 from services.image_service import image_version_diff
-from services.helm_service import diff, apply_upgrade
+from services.helm_service import diff, apply_upgrade, plan_upgrade
 from services.metadata_service import set_ownership_metadata
 from services.pod_label_service import rolling_update_pod_labels
 
@@ -51,6 +51,12 @@ if __name__ == '__main__':
 
     if args.action == 'show-default-config':
         print_default_config()
+    elif args.action == 'plan':
+        plan_upgrade(chart_path=args.chart,
+             release_name=args.release_name,
+             values=args.values,
+             config_path=args.config,
+             selector=args.selector)
     elif args.action == 'apply':
         apply_upgrade(chart_path=args.chart,
              release_name=args.release_name,